Extract target info
