+++ src/cargo/ops/cargo_rustc/context/mod.rs
@@ -5,13 +5,12 @@ use std::env;
 use std::path::{Path, PathBuf};
 use std::str::{self, FromStr};
 use std::sync::Arc;
-use std::cell::RefCell;
 
 use jobserver::Client;
 
 use core::{Package, PackageId, PackageSet, Profile, Resolve, Target};
 use core::{Dependency, Profiles, Workspace};
-use util::{internal, profile, Cfg, CfgExpr, Config, ProcessBuilder};
+use util::{internal, profile, Cfg, CfgExpr, Config};
 use util::errors::{CargoResult, CargoResultExt};
 
 use super::TargetConfig;
@@ -27,6 +26,10 @@ use self::unit_dependencies::build_unit_dependencies;
 mod compilation_files;
 use self::compilation_files::CompilationFiles;
 pub use self::compilation_files::Metadata;
+
+mod target_info;
+pub use self::target_info::TargetFileType;
+use self::target_info::TargetInfo;
 
 /// All information needed to define a Unit.
 ///
@@ -63,17 +66,6 @@ pub struct Unit<'a> {
     pub kind: Kind,
 }
 
-/// Type of each file generated by a Unit.
-#[derive(Copy, Clone, PartialEq, Eq, Debug)]
-pub enum TargetFileType {
-    /// Not a special file type.
-    Normal,
-    /// It is something you can link against (e.g. a library)
-    Linkable,
-    /// It is a piece of external debug information (e.g. *.dSYM and *.pdb)
-    DebugInfo,
-}
-
 /// The build context, containing all information about a build task
 pub struct Context<'a, 'cfg: 'a> {
     /// The workspace the build is for
@@ -103,34 +95,6 @@ pub struct Context<'a, 'cfg: 'a> {
 
     unit_dependencies: HashMap<Unit<'a>, Vec<Unit<'a>>>,
     files: Option<CompilationFiles<'a, 'cfg>>,
-}
-
-#[derive(Clone, Default)]
-struct TargetInfo {
-    crate_type_process: Option<ProcessBuilder>,
-    crate_types: RefCell<HashMap<String, Option<(String, String)>>>,
-    cfg: Option<Vec<Cfg>>,
-    sysroot_libdir: Option<PathBuf>,
-}
-
-impl TargetInfo {
-    fn discover_crate_type(&self, crate_type: &str) -> CargoResult<Option<(String, String)>> {
-        let mut process = self.crate_type_process.clone().unwrap();
-
-        process.arg("--crate-type").arg(crate_type);
-
-        let output = process.exec_with_output().chain_err(|| {
-            format!(
-                "failed to run `rustc` to learn about \
-                 crate-type {} information",
-                crate_type
-            )
-        })?;
-
-        let error = str::from_utf8(&output.stderr).unwrap();
-        let output = str::from_utf8(&output.stdout).unwrap();
-        Ok(parse_crate_type(crate_type, error, &mut output.lines())?)
-    }
 }
 
 impl<'a, 'cfg> Context<'a, 'cfg> {
@@ -238,107 +202,15 @@ impl<'a, 'cfg> Context<'a, 'cfg> {
             _ => true,
         };
 
-        if host_target_same {
-            let info = self.probe_target_info_kind(Kind::Target)?;
-            self.host_info = info.clone();
-            self.target_info = info;
+        self.host_info = TargetInfo::new(self, Kind::Host)?;
+        self.target_info = if host_target_same {
+            self.host_info.clone()
         } else {
-            self.host_info = self.probe_target_info_kind(Kind::Host)?;
-            self.target_info = self.probe_target_info_kind(Kind::Target)?;
-        }
+            TargetInfo::new(self, Kind::Target)?
+        };
         self.compilation.host_dylib_path = self.host_info.sysroot_libdir.clone();
         self.compilation.target_dylib_path = self.target_info.sysroot_libdir.clone();
         Ok(())
-    }
-
-    fn probe_target_info_kind(&self, kind: Kind) -> CargoResult<TargetInfo> {
-        let rustflags = env_args(
-            self.config,
-            &self.build_config,
-            self.info(&kind),
-            kind,
-            "RUSTFLAGS",
-        )?;
-        let mut process = self.config.rustc()?.process();
-        process
-            .arg("-")
-            .arg("--crate-name")
-            .arg("___")
-            .arg("--print=file-names")
-            .args(&rustflags)
-            .env_remove("RUST_LOG");
-
-        if kind == Kind::Target {
-            process.arg("--target").arg(&self.target_triple());
-        }
-
-        let crate_type_process = process.clone();
-        const KNOWN_CRATE_TYPES: &[&str] =
-            &["bin", "rlib", "dylib", "cdylib", "staticlib", "proc-macro"];
-        for crate_type in KNOWN_CRATE_TYPES.iter() {
-            process.arg("--crate-type").arg(crate_type);
-        }
-
-        let mut with_cfg = process.clone();
-        with_cfg.arg("--print=sysroot");
-        with_cfg.arg("--print=cfg");
-
-        let mut has_cfg_and_sysroot = true;
-        let output = with_cfg
-            .exec_with_output()
-            .or_else(|_| {
-                has_cfg_and_sysroot = false;
-                process.exec_with_output()
-            })
-            .chain_err(|| "failed to run `rustc` to learn about target-specific information")?;
-
-        let error = str::from_utf8(&output.stderr).unwrap();
-        let output = str::from_utf8(&output.stdout).unwrap();
-        let mut lines = output.lines();
-        let mut map = HashMap::new();
-        for crate_type in KNOWN_CRATE_TYPES {
-            let out = parse_crate_type(crate_type, error, &mut lines)?;
-            map.insert(crate_type.to_string(), out);
-        }
-
-        let mut sysroot_libdir = None;
-        if has_cfg_and_sysroot {
-            let line = match lines.next() {
-                Some(line) => line,
-                None => bail!(
-                    "output of --print=sysroot missing when learning about \
-                     target-specific information from rustc"
-                ),
-            };
-            let mut rustlib = PathBuf::from(line);
-            if kind == Kind::Host {
-                if cfg!(windows) {
-                    rustlib.push("bin");
-                } else {
-                    rustlib.push("lib");
-                }
-                sysroot_libdir = Some(rustlib);
-            } else {
-                rustlib.push("lib");
-                rustlib.push("rustlib");
-                rustlib.push(self.target_triple());
-                rustlib.push("lib");
-                sysroot_libdir = Some(rustlib);
-            }
-        }
-
-        let cfg = if has_cfg_and_sysroot {
-            Some(lines.map(Cfg::from_str).collect::<CargoResult<_>>()?)
-        } else {
-            None
-        };
-
-        Ok(TargetInfo {
-            crate_type_process: Some(crate_type_process),
-            crate_types: RefCell::new(map),
-            cfg,
-            sysroot_libdir,
-        })
     }
 
     /// Builds up the `used_in_plugin` internal to this context from the list of
@@ -707,45 +579,4 @@ fn env_args(
     }
 
     Ok(Vec::new())
-}
-
-/// Takes rustc output (using specialized command line args), and calculates the file prefix and
-/// suffix for the given crate type, or returns None if the type is not supported. (e.g. for a
-/// rust library like libcargo.rlib, prefix = "lib", suffix = "rlib").
-///
-/// The caller needs to ensure that the lines object is at the correct line for the given crate
-/// type: this is not checked.
-// This function can not handle more than 1 file per type (with wasm32-unknown-emscripten, there
-// are 2 files for bin (.wasm and .js))
-fn parse_crate_type(
-    crate_type: &str,
-    error: &str,
-    lines: &mut str::Lines,
-) -> CargoResult<Option<(String, String)>> {
-    let not_supported = error.lines().any(|line| {
-        (line.contains("unsupported crate type") || line.contains("unknown crate type"))
-            && line.contains(crate_type)
-    });
-    if not_supported {
-        return Ok(None);
-    }
-    let line = match lines.next() {
-        Some(line) => line,
-        None => bail!(
-            "malformed output when learning about \
-             crate-type {} information",
-            crate_type
-        ),
-    };
-    let mut parts = line.trim().split("___");
-    let prefix = parts.next().unwrap();
-    let suffix = match parts.next() {
-        Some(part) => part,
-        None => bail!(
-            "output of --print=file-names has changed in \
-             the compiler, cannot parse"
-        ),
-    };
-
-    Ok(Some((prefix.to_string(), suffix.to_string())))
 }
